+++ projects/Win/vc141/cds.vcxproj.filters
@@ -1261,5 +1261,8 @@
     <ClInclude Include="..\..\..\cds\details\size_t_cast.h">
       <Filter>Header Files\cds\details</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\..\cds\container\concurrent_hopscotch_hash_set.h">
+      <Filter>Header Files\cds\container</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
